+++ NEWS.md
@@ -13,11 +13,13 @@
   - `generate_all_months()`: Generate all months with `yyyymm` format in a year
   - `get_url()`: Put together the data file url
   - `download_flights_data()`: Download and read ANAC flight data
+- Three separate vignettes. A general intro to the package, and more detailed vignettes on `read_flighs` and `read_airports()`.
 
 * Minor changes:
 - new parameter `select` in `read_flighs()`, allowing the user to specify the columns that should be read.
 - new tests of `read_flighs()`. Coverage of 95.24%. [Closes #5](https://github.com/ipeaGIT/flightsbr/issues/5).
 - New checks on `date` input. [Closes #2](https://github.com/ipeaGIT/flightsbr/issues/2).
+- Functions now should fail gracefully in case of problems with internet connection. [Closes #7](https://github.com/ipeaGIT/flightsbr/issues/7).
 
 
 -------------------------------------------------------